% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_annoy.R
\name{knn.annoy}
\alias{knn.annoy}
\title{K Nearest Neighbour Search}
\usage{
knn.annoy(data, K = 30, trees = 150, return_dist = FALSE)
}
\arguments{
\item{data}{matrix; input data matrix}

\item{K}{integer; number of nearest neighbours}

\item{trees}{integer; number of trees in annoylib; more trees - more precise result}

\item{return_dist}{bool; if TRUE a list of indices and distances is returned}
}
\value{
a n-by-k matrix of neighbors indices or list(IND,DIST) see parameter return_dist
}
\description{
Uses a annoylib to find K nearest neighbors in dataset
see https://github.com/spotify/annoy/
}
\details{
If used as precomputed matrix for Rphenoannoy, discard first column - 1st nearest 
neigbor is the point itself
}
\examples{
iris_unique <- unique(iris) # Remove duplicates
data <- as.matrix(iris_unique[,1:4])
neighbors <- knn.annoy(data, k=10)

}
