% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_balltree.R
\name{knn.balltree}
\alias{knn.balltree}
\title{K Nearest Neighbour Search}
\usage{
knn.balltree(X, K = 30)
}
\arguments{
\item{K}{integer; number of nearest neighbours}

\item{data}{matrix; input data matrix}
}
\value{
a n-by-k matrix of neighbors indices or list(IND,DIST)
}
\description{
Uses a paralell version of vantage point tree to find K nearest neighbors in dataset
see https://lvdmaaten.github.io/tsne/
}
\details{
If used as precomputed matrix for Rphenoannoy, discard first column - 1st nearest 
neigbor is the point itself
}
\examples{
iris_unique <- unique(iris) # Remove duplicates
data <- as.matrix(iris_unique[,1:4])
neighbors <- knn.balltree(data, k=10)

}
