% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph.R
\name{Rphenoannoy}
\alias{Rphenoannoy}
\title{RphenoGraph clustering}
\usage{
Rphenoannoy(
  data = NULL,
  k = 30,
  neighborMatrix = NULL,
  method = "annoy",
  trees = 150,
  threshold = 0L
)
}
\arguments{
\item{data}{matrix; input data matrix}

\item{k}{integer; number of nearest neighbours (default:30)}

\item{neighborMatrix}{matrix; (optional) precomputed indices of nearest neigbors;1-based.}

\item{method}{character; if =="annoy" approximate nearest neigbors are computed (default:"annoy"), 
"default" - the RANN::nn2 function is used, "balltree" parallelized version of vantage point tree is used}

\item{trees}{int; number of trees for annoylib, more trees more precise but slower (default:150)}

\item{threshold}{int; minimum number of common neighbors to add edge into graph}
}
\value{
a list contains the KNN matrix and a communities object, the operations of this class contains:
\item{print}{returns the communities object itself, invisibly.}
\item{length}{returns an integer scalar.}
\item{sizes}{returns a numeric vector.}
\item{membership}{returns a numeric vector, one number for each vertex in the graph that was the input of the community detection.}
\item{modularity}{returns a numeric scalar.}
\item{algorithm}{returns a character scalar.}
\item{crossing}{returns a logical vector.}
\item{is_hierarchical}{returns a logical scalar.}
\item{merges}{returns a two-column numeric matrix.}
\item{cut_at}{returns a numeric vector, the membership vector of the vertices.}
\item{as.dendrogram}{returns a dendrogram object.}
\item{show_trace}{returns a character vector.}
\item{code_len}{returns a numeric scalar for communities found with the InfoMAP method and NULL for other methods.}
\item{plot}{for communities objects returns NULL, invisibly.}
}
\description{
R implementation of the PhenoGraph algorithm
}
\details{
A simple R implementation of the [PhenoGraph](http://www.cell.com/cell/abstract/S0092-8674(15)00637-6) algorithm,
which is a clustering method designed for high-dimensional single-cell data analysis. It works by creating a graph ("network") representing
phenotypic similarities between cells by calclating the Jaccard coefficient between nearest-neighbor sets, and then identifying communities
using the well known [Louvain method](https://sites.google.com/site/findcommunities/) in this graph.
}
\examples{
iris_unique <- unique(iris) # Remove duplicates
data <- as.matrix(iris_unique[,1:4])
Rphenograph_out <- Rphenoannoy(data, k = 45)
modularity(Rphenograph_out[[2]])
membership(Rphenograph_out[[2]])
iris_unique$phenograph_cluster <- factor(membership(Rphenograph_out[[2]]))
ggplot(iris_unique, aes(x=Sepal.Length, y=Sepal.Width, col=Species, shape=phenograph_cluster)) + geom_point(size = 3)+theme_bw()

}
\references{
Jacob H. Levine and et.al. Data-Driven Phenotypic Dissection of AML Reveals Progenitor-like Cells that Correlate with Prognosis. Cell, 2015.
}
